# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: User Management APIs
host: 'localhost:10004'
basePath: /rbac
schemes:
- http
consumes:
- multipart/form-data
- application/json
produces:
- application/json
- text/plain
paths:
...
  /app:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of User
      operationId: AppList
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: page
        in: query
        type: integer
        description: Page number of the request
      - name: count
        in: query
        type: integer
        description: Number of results per page
      - name: filter
        in: query
        type: string
        description: Filter results based on certain fields
      - name: select
        in: query
        type: string
        description: Comma seperated fields to be displayed
      - name: sort
        in: query
        type: string
        description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error
    post:
      description: Create a new Domian
      operationId: AppCreate
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: data
        in: body
        description: Payload to create a App
        schema:
          $ref: '#/definitions/Appcreate'
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error
...
  /role:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of Roles
      operationId: RolesList
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: page
        in: query
        type: integer
        description: Page number of the request
      - name: count
        in: query
        type: integer
        description: Number of results per page
      - name: filter
        in: query
        type: string
        description: Filter results based on certain fields
      - name: select
        in: query
        type: string
        description: Comma seperated fields to be displayed
      - name: sort
        in: query
        type: string
        description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error
    post:
      description: Create a new role
      operationId: RolesCreate
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: data
        in: body
        description: Payload to create a role
        schema:
          $ref: '#/definitions/Rolescreate'
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error
...
definitions:
...
  Appcreate:
    properties:
      type:
        type: string
        enum:
        - Management
        - Distribution
...
  Rolescreate:
    required:
    - entity
    - app
    properties:
      description:
        type: string
      app:
        type: string
      entity:
        type: string
      entityName:
        type: string
      fields:
        type: [string, 'null']
...